Refactor: extract methods, constants, and remove dead code across services
Improve readability by extracting long methods into smaller focused ones,
replacing magic strings with constants, removing unused method overloads
(heartbeat callback, parseUnmappedResult), and simplifying emoji regex.

+++ backend/src/main/java/com/example/rag/model/TokenRecordingChatModel.java
@@ -42,8 +42,8 @@ public class TokenRecordingChatModel implements ChatModel {
             var usage = response.getMetadata().getUsage();
             if (usage == null) return;
 
-            int inputTokens = (int) usage.getPromptTokens();
-            int outputTokens = (int) usage.getCompletionTokens();
+            int inputTokens = usage.getPromptTokens();
+            int outputTokens = usage.getCompletionTokens();
             if (inputTokens == 0 && outputTokens == 0) return;
 
             java.util.UUID userId = resolveCurrentUserId();
